Refactor storage
