+++ burn-compute/src/storage/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod bytes_cpu;
+
+pub use base::*;
+pub use bytes_cpu::*;
